Update plugins only if there is a newer version
A 'git pull' always generates a change event in Puppet even if there are
no changes in the upstream repository. This is bad for monitoring, for
example the puppet dashboard always shows all the plugins installed by
rbenv-puppet as changes for every run on every node.

This change makes sure that 'git pull' is only run when there is a new
version of the plugin upstream.